Rework registration of rustdoc provider
