+++ crates/assistant/src/assistant.rs
@@ -20,7 +20,7 @@ pub(crate) use completion_provider::*;
 pub(crate) use context_store::*;
 use fs::Fs;
 use gpui::{actions, AppContext, Global, SharedString, UpdateGlobal};
-use indexed_docs::{IndexedDocsRegistry, Provider};
+use indexed_docs::IndexedDocsRegistry;
 pub(crate) use inline_assistant::*;
 pub(crate) use model_selector::*;
 use semantic_index::{CloudEmbeddingProvider, SemanticIndex};
@@ -293,7 +293,6 @@ pub fn init(fs: Arc<dyn Fs>, client: Arc<Client>, cx: &mut AppContext) {
     inline_assistant::init(fs.clone(), client.telemetry().clone(), cx);
     terminal_inline_assistant::init(fs.clone(), client.telemetry().clone(), cx);
     IndexedDocsRegistry::init_global(cx);
-    register_indexed_docs_providers(cx);
 
     CommandPaletteFilter::update_global(cx, |filter, _cx| {
         filter.hide_namespace(Assistant::NAMESPACE);
@@ -326,12 +325,6 @@ fn register_slash_commands(cx: &mut AppContext) {
     slash_command_registry.register_command(diagnostics_command::DiagnosticsCommand, true);
     slash_command_registry.register_command(rustdoc_command::RustdocSlashCommand, false);
     slash_command_registry.register_command(fetch_command::FetchSlashCommand, false);
-}
-
-fn register_indexed_docs_providers(cx: &mut AppContext) {
-    let indexed_docs_registry = IndexedDocsRegistry::global(cx);
-
-    indexed_docs_registry.register_provider(Provider::rustdoc());
 }
 
 pub fn humanize_token_count(count: usize) -> String {
