+++ crates/extension/src/extension_docs_indexer.rs
@@ -3,25 +3,21 @@ use std::sync::Arc;
 use anyhow::{anyhow, Result};
 use async_trait::async_trait;
 use futures::FutureExt;
-use indexed_docs::{IndexDocs, IndexedDocsDatabase, IndexedDocsProvider, PackageName};
+use indexed_docs::{IndexDocs, IndexedDocsDatabase, PackageName};
 use wasmtime_wasi::WasiView;
 
 use crate::wasm_host::{WasmExtension, WasmHost};
 
 pub struct ExtensionDocsIndexer {
     pub(crate) extension: WasmExtension,
+    #[allow(unused)]
     pub(crate) host: Arc<WasmHost>,
     pub(crate) name: Arc<str>,
 }
 
 #[async_trait]
 impl IndexDocs for ExtensionDocsIndexer {
-    async fn index(
-        &self,
-        package: PackageName,
-        database: Arc<IndexedDocsDatabase>,
-        provider: Box<dyn IndexedDocsProvider + Send + Sync + 'static>,
-    ) -> Result<()> {
+    async fn index(&self, package: PackageName, database: Arc<IndexedDocsDatabase>) -> Result<()> {
         self.extension
             .call({
                 let name = self.name.clone();
