+++ crates/indexed_docs/src/indexed_docs.rs
@@ -1,9 +1,7 @@
-mod indexer;
 mod providers;
 mod registry;
 mod store;
 
-pub use crate::indexer::{IndexDocs, IndexedDocsProvider};
 pub use crate::providers::rustdoc::*;
 pub use crate::registry::*;
 pub use crate::store::*;
